set Command.Type in Command subclass constructors, not as argument. move CornExceptionMappers.
