+++ src/main/java/org/cornfields/simulator/game/CommandFactory.java
@@ -14,17 +14,11 @@ public class CommandFactory {
     }
 
     if (commandString.equals("REGISTER") & argument.isPresent()) {
-      return Optional.of(
-          new RegisterCommand(Command.Type.REGISTER, sourceNumber, argument.get())
-      );
+      return Optional.of(new RegisterCommand(sourceNumber, argument.get()));
     } else if (commandString.equals("TRAVEL") & argument.isPresent()) {
-      return Optional.of(
-          new TravelCommand(Command.Type.TRAVEL, sourceNumber, argument.get())
-      );
+      return Optional.of(new TravelCommand(sourceNumber, argument.get()));
     } else if (commandString.equals("CORN")) {
-      return Optional.of(
-          new Command(Command.Type.CORN, sourceNumber)
-      );
+      return Optional.of(new Command(Command.Type.CORN, sourceNumber));
     }
 
     return Optional.empty();
